+++ changelog.md
@@ -4,6 +4,7 @@
 
 - added: hide controls when zooming in
 - added: long press profile to open account switcher
+- changed: font settings moved to appearance
 
 ## Build 41